Create beta_norm
This is a python script I have written to do the same thing as the established MATLAB code I had in this project already. This code should work on any platform (note, I use python 2.7). The header of this code should explain the use of the code as well as the required inputs. If there are any questions, bugs found, or comments, please let me know at MarkCBabin@gmail.com